Added basic support for syncing a radio stations feed from a server with a local core data database.
Stations are inserted, updated and deleted every minute (timing should obviously be adjusted later on).

The version of the feed is checked before the DB is synced. Currently the station name is used as the key, but this needs to be revised. User-defined and cloud-defined stations are separated, so a sync shouldn't nuke local stations.